Added another selector.
The selector that was just added is actually not working and hasn't been added as an id. What I am trying to do is something like this:

Home, Favorite Media: Anime.
                                     Media

As you can see, the subsections on that certain page are next to each other in an imaginary box. Though I don't have the current knowledge to do this, so for right now, it will be an issue!!